ancora riorganizzazione file. Creato AdminService per disaccoppiare meglio le funzionalità

+++ backend/src/dao/AdminDAO.ts
@@ -1,4 +1,3 @@
-import { Transaction } from 'sequelize';
 import { User } from '../models/UserModel.js';
 import { UserCreationData } from '../models/UserModel.js';
 import { AppErrorEnum } from '../utils/StatusMessages.js';
@@ -7,7 +6,7 @@ import { ErrorFactory } from '../factory/ErrorFactory.js';
 interface IAdminDAO {
   create(data: UserCreationData): Promise<User>;
   findAll(): Promise<User[]>;
-  findById(user_id: number, t?: Transaction): Promise<User | null>;
+  findById(user_id: number): Promise<User | null>;
   update(user_id: number, data: Partial<UserCreationData>): Promise<number>;
   delete(user_id: number): Promise<number>; 
 }
@@ -23,11 +22,8 @@ export class AdminDAO implements IAdminDAO {
     
   }
 
-  async findById(user_id: number, t?: Transaction): Promise<User | null> {
+  async findById(user_id: number): Promise<User | null> {
     try {
-      if (t) {
-        return await User.findByPk(user_id, { transaction: t });
-      }
       return await User.findByPk(user_id);
     } catch (err) {
       throw ErrorFactory.getError(AppErrorEnum.USER_NOT_FOUND);
@@ -44,17 +40,17 @@ export class AdminDAO implements IAdminDAO {
 
   async findByEmail(email: string): Promise<User | null> {
     // Siccome a noi serve che non trova l'email, non possiamo mettere l'errore metterlo
-      return await User.findOne({ where: { email } });
+      return await User.findOne({ where: { email: email } });
   }
 
   async findByUsername(username: string): Promise<User | null> {
     // Stessa cosa per l'username
-      return await User.findOne({ where: { username } });
+      return await User.findOne({ where: { username: username } });
   }
 
   async update(user_id: number, data: Partial<UserCreationData>): Promise<number> {
     try{
-      const [affectedCount] = await User.update(data, { where: { user_id } });
+      const [affectedCount] = await User.update(data, { where: { user_id: user_id } });
       return affectedCount;
     } catch (err){
       throw ErrorFactory.getError(AppErrorEnum.INTERNAL_ERROR);
@@ -63,7 +59,7 @@ export class AdminDAO implements IAdminDAO {
 
   async delete(user_id: number): Promise<number> {
     try{
-      return await User.destroy({ where: { user_id } });
+      return await User.destroy({ where: { user_id: user_id } });
     } catch (err){
       throw ErrorFactory.getError(AppErrorEnum.USER_NOT_FOUND);
     }
